+++ app/src/main/java/org/librelab/marklibre/ColorButton.kt
@@ -7,7 +7,7 @@ import android.graphics.Paint
 import android.util.AttributeSet
 import android.view.View
 
-class ColorButton @JvmOverloads constructor(
+open class ColorButton @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null
 ) : View(context, attrs) {